+++ frontend/src/lib/utils.ts
@@ -1504,6 +1504,8 @@ export type Item = {
 	extra?: Snippet
 	id?: string
 	tooltip?: string
+	separatorTop?: boolean
+	submenuItems?: Item[]
 }
 
 export function isObjectTooBig(obj: any): boolean {